exercises/01-ml/04-recommender-systems/requirements.txt
# Core ML Libraries
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Advanced ML
xgboost>=2.0.0,<3.0.0
optuna>=3.3.0,<4.0.0

# Recommender Systems
scipy>=1.11.0,<2.0.0
tensorflow>=2.13.0,<3.0.0
faiss-cpu>=1.7.4,<2.0.0

# Model Persistence & Tracking
joblib>=1.3.0,<2.0.0
mlflow>=2.8.0,<3.0.0

# API & Production Serving
flask>=3.0.0,<4.0.0
gunicorn>=21.2.0,<22.0.0
pydantic>=2.4.0,<3.0.0

# Monitoring
prometheus-client>=0.18.0,<1.0.0

# Configuration
pyyaml>=6.0,<7.0.0

# Development & Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
black>=23.7.0,<24.0.0
flake8>=6.1.0,<7.0.0
